requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
alembic
httpx
python-dotenv
cryptography
jinja2
py-ulid
passlib[bcrypt]